Fixed empty categories bug, cleaned some code.
@mosaed_otaibi

I fixed the bug of the empty category, that is when there
are no categories the user wil get an error when trying to
add a task (because there no categories to add the task to).
I fixed it by preventing the user from deleting the last
category, so when a category is the last, the deletion opt-
-ion will be removed from the category pop-up button. I re-
-moved man useless print statements, and added indications
of where each error catch is occurring (by printing where).